+++ src/constants.ts
@@ -1,2 +1,4 @@
 export const PLUGIN_NAME = "markdown-pages";
 export const THEME_NAME = PLUGIN_NAME;
+export const DEFAULT_PAGES_LABEL = "Pages";
+export const DEFAULT_OUTPUT_DIR_NAME = "pages";