+++ src/models/markdown-page-collection.ts
@@ -1,10 +1,11 @@
 import { IMarkdownPageCollection } from "../interfaces/markdown-page-collection-interface";
 import { IMarkdownPage } from "../interfaces/markdown-page-interface";
-import { getFileExtension, getFileName, getHumanReadableNameFromFileName } from "../utilities/path-utilities";
+import { getDirectoryName, getFileExtension, getFileName, getHumanReadableNameFromFileName } from "../utilities/path-utilities";
 import { MarkdownPage } from "./markdown-page";
-import { join } from "path";
+import { join, relative } from "path";
 import { resolve } from "url";
 import { getDirectoryContents, isDirectory } from "../utilities/filesystem-utilities";
+import { Logger } from "typedoc/dist/lib/utils/loggers";
 
 export class MarkdownPageCollection implements IMarkdownPageCollection, IMarkdownPage {
 	public title: string;
@@ -14,10 +15,17 @@ export class MarkdownPageCollection implements IMarkdownPageCollection, IMarkdow
 	public url: string;
 	public contents: string;
 
-	constructor(path: string, url: string) {
+	private _logger: Logger;
+
+	constructor(logger: Logger, path: string, url: string) {
+		this._logger = logger;
 		this.path = path;
 		this.url = url;
 		this.children = [];
+
+		if (!isDirectory(path)) {
+			throw new Error(`Markdown page collection path "${path}" is not a directory.`);
+		}
 
 		this._parseTitle(path);
 	}
@@ -29,10 +37,11 @@ export class MarkdownPageCollection implements IMarkdownPageCollection, IMarkdow
 			for (const childItem of childItems) {
 				let childPage: IMarkdownPage;
 				const childPath = join(this.path, childItem)
-				const childUrl = resolve(this.url, this._getItemOutputFilename(childItem));
+				const isChildADirectory = isDirectory(childPath);
+				const childUrl = this._getItemUrl(childItem, isChildADirectory);
 
-				if (isDirectory(childPath)) {
-					childPage = new MarkdownPageCollection(childPath, childUrl);
+				if (isChildADirectory) {
+					childPage = new MarkdownPageCollection(this._logger, childPath, childUrl);
 				} else {
 					childPage = new MarkdownPage(childPath, childUrl);
 				}
@@ -44,27 +53,80 @@ export class MarkdownPageCollection implements IMarkdownPageCollection, IMarkdow
 				} else {
 					this.children.push(childPage);
 				}
-
-
 			}
 
-			// Add option for creating an index file
+			if (!this.contents) {
+				this._generateIndexContents();
+			}
 		} catch (e) {
 			throw new Error(`Failed to read page collection contents. ${e}`);
 		}
 	}
 
+	public log(): void {
+		this._recursiveLog(this as IMarkdownPage, 0);
+	}
+
 	private _parseTitle(path: string): void {
-		const fileName = getFileName(path);
-		const fileExtension = getFileExtension(fileName);
-		const itemNameWithoutExtension = fileName.slice(0, (1 + fileExtension.length) * -1);
-		const humanReadableName = getHumanReadableNameFromFileName(itemNameWithoutExtension);
+		const dirName = getFileName(path);
+		const humanReadableName = getHumanReadableNameFromFileName(dirName);
 		this.title = humanReadableName;
 	}
 
-	private _getItemOutputFilename(sourceItemName: string): string { // TODO: rename this
-		const fileExtension = getFileExtension(sourceItemName);
-		const itemNameWithoutExtension = sourceItemName.slice(0, (1 + fileExtension.length) * -1);
-		return itemNameWithoutExtension + ".html";
+	private _getItemUrl(sourceItemName: string, isDirectory: boolean): string {
+		const urlPathToCollection = getDirectoryName(this.url)
+			.replace(/\/?$/, "/"); // ensure last character is a slash
+
+		if (isDirectory) {
+			const newDir = resolve(urlPathToCollection, sourceItemName)
+				.replace(/\/?$/, "/"); // ensure last character is a slash
+			return resolve(newDir, "index.html");
+		} else {
+			const fileExtension = getFileExtension(sourceItemName);
+			const sourceItemNameWithoutExtension = sourceItemName.slice(0, (1 + fileExtension.length) * -1);
+			return resolve(urlPathToCollection, sourceItemNameWithoutExtension) + ".html";
+		}
+	}
+
+	private _recursiveLog(page: IMarkdownPage, depth: number): void {
+		const isCollection = !!(page as any).children;
+
+		if (depth === 0) {
+			this._logger.verbose(page.title);
+		} else {
+			let prefix = "";
+			for (let i = 0; i < depth; i++) {
+				prefix += " ";
+			}
+			prefix += "|__"
+			this._logger.verbose(`${prefix}${page.title}`);
+		}
+
+		if (isCollection) {
+			for (const child of (page as IMarkdownPageCollection).children) {
+				const newDepth = depth + Math.floor(page.title.length / 2) + (depth === 0 ? 0 : 3);
+				this._recursiveLog(child, newDepth);
+			}
+		}
+	}
+
+	private _generateIndexContents(): void {
+		try {
+			this._logger.verbose(`Generating index contents for page "${this.title}"...`);
+
+			let contents = "## Contents";
+
+			for (const child of this.children) {
+				const childUrl = getFileName(child.url);
+				contents += `\n- [${child.title}](${childUrl})`;
+			}
+
+			this.contents = contents;
+			this._logger.verbose(this.contents);
+		} catch (e) {
+			const errorMessage = `Failed to generate index contents. ${e}`;
+			this._logger.error(errorMessage);
+			throw new Error(errorMessage);
+		}
 	}
 }