+++ CHANGELOG.md
@@ -3,9 +3,26 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.2.0] - 2019-04-23
+### Added
+- `mdPagesLabel` option for specifying the top-level page title
+- Improved navigation handling and breadcrumbs
+- Better logging for discovered pages
+- Improved README usage documentation
+- Support for customizable index pages
+- Support for auto-generated index pages
+- Improved HTML file naming in output
+- Allow theme to be used by specifying plugin name to --theme option
+- `mdPagesOutputDirName` option for specifying the name of the output directory
+
+### Changed
+- Changed `mdPagesSourceDir` option to be named `mdPagesSourcePath` for clarity
+
 ## [0.1.0] - 2019-04-14
 ### Added
 - Initial commit of the plugin with README
 - This CHANGELOG file.
 
+[Unreleased]: https://github.com/mipatterson/typedoc-plugin-markdown-pages/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/mipatterson/typedoc-plugin-markdown-pages/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/mipatterson/typedoc-plugin-markdown-pages/releases/tag/v0.1.0